part C/requirements.txt
# requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3

# MongoDB dependencies
pymongo==4.6.3
dnspython==2.6.1  # For MongoDB SRV connection strings

# Environment management
python-dotenv==1.0.0

# Request handling
Flask-Cors==4.0.0

# Form handling and validation
WTForms==3.0.1

# Date and time handling
pytz==2023.3

# Additional utilities
blinker==1.6.2